requirements.txt
discord.py
python-dotenv
flask
pytz
yt_dlp
asyncio
pynacl
